# src/main/resources/contract/api.yaml
openapi: 3.0.0
info:
  title: Proyecto DAW - Empresa de Tatuajes
  description: API para manejar el back-end de una empresa de tatuajes.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Servidor local
paths:
...
  /Orders:
    get:
      summary: Obtener todos los pedidos
      description: "Permite a cualquier usuario obtener la lista de pedidos realizados."
      operationId: getOrders
      tags:
        - Orders
      responses:
        "200":
          description: Lista de pedidos obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                title: OrdersResponse
                properties:
                  orders:
                    type: array
                    description: Lista de pedidos realizados
                    items:
                      $ref: "#/components/schemas/Orders"
        "404":
          description: No se encontraron pedidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Crear un nuevo pedido
      description: "Permite a un usuario realizar un nuevo pedido. El pedido es creado con la información proporcionada."
      operationId: createOrder
      tags:
        - Orders
      requestBody:
        description: Datos del nuevo pedido
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Orders"
      responses:
        "201":
          description: Pedido creado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        "400":
          description: Solicitud incorrecta (Bad Request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /Orders/{idOrder}:
    get:
      summary: Obtener un pedido por ID
      description: "Permite a un usuario obtener los detalles de un pedido específico mediante su ID único."
      operationId: getOrderById
      tags:
        - Orders
      parameters:
        - name: idOrder
          in: path
          required: true
          description: ID único del pedido a obtener
          schema:
            type: integer
      responses:
        "200":
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        "404":
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Actualizar un pedido por ID
      description: "Permite a un usuario actualizar los datos de un pedido específico mediante su ID único."
      operationId: updateOrder
      tags:
        - Orders
      parameters:
        - name: idOrder
          in: path
          required: true
          description: ID único del pedido a actualizar
          schema:
            type: integer
      requestBody:
        description: Nuevos datos del pedido
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Orders"
      responses:
        "200":
          description: Pedido actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Orders"
        "400":
          description: Solicitud incorrecta (Bad Request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Eliminar un pedido por ID
      description: "Permite a un usuario eliminar un pedido específico mediante su ID único."
      operationId: deleteOrder
      tags:
        - Orders
      parameters:
        - name: idOrder
          in: path
          required: true
          description: ID único del pedido a eliminar
          schema:
            type: integer
      responses:
        "204":
          description: Pedido eliminado exitosamente
        "404":
          description: Pedido no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Orders:
      type: object
      properties:
        idOrder:
          type: integer
          format: int32
          description: Identificador único del pedido
        idUser:
          type: integer
          format: int32
          description: Identificador único del usuario que realizó el pedido
        totalQuantity:
          type: integer
          format: int32
          description: Cantidad total de productos en el pedido
        totalPrice:
          type: number
          format: double
          description: Precio total del pedido
        date:
          type: string
          format: date-time
          description: Fecha
        orderStatus:
          $ref: '#/components/schemas/OrderStatus'
      required:
        - idOrder
        - idUser
        - totalQuantity
        - totalPrice
        - date
        - orderStatus

    # Estado de la orden (correcto - no tocar)
    OrderStatus:
      type: string
      enum: ['PENDIENTE', 'PAGADO', 'CANCELADO', 'CONFIRMADO']
      default: 'PENDIENTE'

    # Correcto (se supone)
...
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          description: Código del error
        errorMessage:
          type: string
          description: Descripción del error
security:
  - bearerAuth: []